requirements.txt
bcrypt==4.1.2
Flask==2.3.3
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.3
SQLAlchemy==2.0.29
flask_cors==4.0.0
flask-restful==0.3.10